Refactor Application Layer to enforce strict interface-only policy
Moves all concrete service implementations (e.g. `*Service.cs`, `*Processor.cs`, `*Executor.cs`) out of `Valora.Application/Services` and relocates them to `Valora.Infrastructure/Services/AppServices`.

Updates namespaces to align with `Valora.Infrastructure.Services.AppServices` and modifies over 25 unit and integration test files to reflect the new paths. Also migrates the Dependency Injection configuration for these services from the Application layer to the Infrastructure layer, ensuring the Application layer solely defines interfaces and abstract types.
