+++ backend/Valora.Application/DependencyInjection.cs
@@ -1,7 +1,5 @@
 using Microsoft.Extensions.DependencyInjection;
 using Valora.Application.Common.Interfaces;
-using Valora.Application.Services;
-using Valora.Application.Services.BatchJobs;
 
 namespace Valora.Application;
 
@@ -9,25 +7,8 @@ public static class DependencyInjection
 {
     public static IServiceCollection AddApplication(this IServiceCollection services)
     {
-        services.AddScoped<IAuthService, AuthService>();
-        services.AddScoped<IAdminService, AdminService>();
-        services.AddScoped<INotificationService, NotificationService>();
-        services.AddScoped<IContextReportService, ContextReportService>();
-        services.AddScoped<IContextAnalysisService, ContextAnalysisService>();
-        services.AddScoped<IUserAiProfileService, UserAiProfileService>();
-        services.AddScoped<IContextDataProvider, ContextDataProvider>();
-        services.AddScoped<IMapService, MapService>();
-        services.AddScoped<IBatchJobService, BatchJobService>();
-        services.AddScoped<IBatchJobExecutor, BatchJobExecutor>();
-        services.AddScoped<IWorkspaceService, WorkspaceService>();
-        services.AddScoped<IWorkspaceMemberService, WorkspaceMemberService>();
-        services.AddScoped<IWorkspacePropertyService, WorkspacePropertyService>();
-
-        // Batch Job Processors
-        services.AddScoped<IBatchJobProcessor, CityIngestionJobProcessor>();
-        services.AddScoped<IBatchJobProcessor, MapGenerationJobProcessor>();
-        services.AddScoped<IBatchJobProcessor, AllCitiesIngestionJobProcessor>();
-
+        // Currently empty as all Application services logic has been moved to Infrastructure
+        // Application layer should only declare abstract Interfaces/DTOs.
         return services;
     }
 }
