+++ backend/Valora.Infrastructure/Services/AppServices/BatchJobs/AllCitiesIngestionJobProcessor.cs
@@ -4,7 +4,7 @@ using Valora.Domain.Entities;
 using Valora.Domain.Enums;
 using Valora.Domain.Extensions;
 
-namespace Valora.Application.Services.BatchJobs;
+namespace Valora.Infrastructure.Services.AppServices.BatchJobs;
 
 public class AllCitiesIngestionJobProcessor : IBatchJobProcessor
 {
