+++ backend/Valora.UnitTests/Services/ContextAnalysisServiceTests.cs
@@ -1,7 +1,7 @@
 using Moq;
 using Valora.Application.Common.Interfaces;
 using Valora.Application.DTOs;
-using Valora.Application.Services;
+using Valora.Infrastructure.Services.AppServices;
 
 namespace Valora.UnitTests.Services;
