+++ backend/Valora.UnitTests/Services/ContextDataProviderTests.cs
@@ -2,7 +2,7 @@ using Microsoft.Extensions.Logging;
 using Moq;
 using Valora.Application.Common.Interfaces;
 using Valora.Application.DTOs;
-using Valora.Application.Services;
+using Valora.Infrastructure.Services.AppServices;
 
 namespace Valora.UnitTests.Services;
