+++ backend/Valora.UnitTests/Services/ContextReportServiceTests.cs
@@ -5,7 +5,7 @@ using Valora.Application.Common.Exceptions;
 using Valora.Application.Common.Interfaces;
 using Valora.Application.DTOs;
 using Valora.Application.Enrichment;
-using Valora.Application.Services;
+using Valora.Infrastructure.Services.AppServices;
 
 namespace Valora.UnitTests.Services;
