+++ backend/Valora.UnitTests/Services/MapServiceTests.cs
@@ -3,7 +3,7 @@ using Moq;
 using Valora.Application.Common.Exceptions;
 using Valora.Application.Common.Interfaces;
 using Valora.Application.DTOs.Map;
-using Valora.Application.Services;
+using Valora.Infrastructure.Services.AppServices;
 using Xunit;
 using Microsoft.Extensions.Caching.Memory;
 using Microsoft.Extensions.Logging;
