+++ backend/Valora.UnitTests/Services/SecurityFixTests.cs
@@ -6,7 +6,7 @@ using Moq;
 using Valora.Application.Common.Interfaces;
 using Valora.Application.Common.Utilities;
 using Valora.Application.DTOs;
-using Valora.Application.Services;
+using Valora.Infrastructure.Services.AppServices;
 using Valora.Domain.Entities;
 using Valora.Domain.Services;
