+++ backend/Valora.UnitTests/Services/UserAiProfileServiceTests.cs
@@ -2,7 +2,7 @@ using Moq;
 using Valora.Application.Common.Interfaces;
 using Valora.Application.Common.Models;
 using Valora.Application.DTOs;
-using Valora.Application.Services;
+using Valora.Infrastructure.Services.AppServices;
 using Valora.Domain.Entities;
 
 namespace Valora.UnitTests.Services;
