+++ backend/Valora.UnitTests/Services/Utilities/PriceOverlayCalculatorTests.cs
@@ -1,7 +1,7 @@
 using System.Globalization;
 using System.Text.Json;
 using Valora.Application.DTOs.Map;
-using Valora.Application.Services.Utilities;
+using Valora.Infrastructure.Services.AppServices.Utilities;
 using Xunit;
 
 namespace Valora.UnitTests.Services.Utilities;
