+++ backend/Valora.UnitTests/Services/WorkspacePropertyServiceTests.cs
@@ -1,7 +1,7 @@
 using Moq;
 using Valora.Application.Common.Interfaces;
 using Valora.Application.Common.Extensions;
-using Valora.Application.Services;
+using Valora.Infrastructure.Services.AppServices;
 using Valora.Domain.Entities;
 using Xunit;
